ecommerce-kafka-elk/consumer/requirements.txt
kafka-python==2.0.2
elasticsearch==8.11.0
python-dotenv==1.0.0